eliminate conversion/promotion errors from build
-enforced through -Wconversion -Wsign-conversion -Wsign-promo
-add perft() to unit tests

+++ include/dctl/dxp/i_message.hpp
@@ -36,8 +36,8 @@ private:
         bool invariant() const
         {
                 return (
-                        do_header().length() == header_length_ &&
-                        do_body().length() <= max_body_length_
+                        static_cast<int>(do_header().length()) == header_length_ &&
+                        static_cast<int>(do_body().length()) <= max_body_length_
                 );
         }
